+++ logstar_stream/processing_steps/ProcessingStep.py
