+++ requirements.txt
@@ -1,5 +1,4 @@
 argparse
 requests
-pyodbc==4.0.30
 sqlalchemy
 pandas
